% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terravars-package.R
\docType{package}
\name{terravars-package}
\alias{terravars}
\alias{terravars-package}
\title{terravars: Terrain, Hydrological and Canopy Indices for Environmental Analysis}
\description{
Provides fast computation of topographic and canopy-derived spatial predictor variables for statistical microclimate downscaling. Functions include solar index, sky-view factor, topographic wetness index (multiple methods), wind shelter indices, and two-stream radiation for vegetated surfaces. Designed for compatibility with the 'terra' package for raster data handling. Core computations are implemented in C++ via 'Rcpp' for performance on large spatial datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ilyamaclean/terravars}
  \item Report bugs at \url{https://github.com/ilyamaclean/terravars/issues}
}

}
\author{
\strong{Maintainer}: Ilya Maclean \email{i.m.d.maclean@exeter.ac.uk} (\href{https://orcid.org/0000-0001-8030-9136}{ORCID})

}
\keyword{internal}
